\name{xsubset}

\docType{package}

\alias{xsubset}
\alias{xsubset-package}

\title{Matrix computations and statistics: variable subset selection}

\description{
  Package for variable-subset model selection.  Also contains two sample
  datasets, \code{pollute} and \code{house}.
}

\references{

  Hofmann, M. and Gatu, C. and Kontoghiorghes, E. J. (2007).  Efficient
  algorithms for computing the best subset regression models for
  large-scale problems.  Computational Statistics and Data Analysis, 52,
  16-29.

  Gatu, C. and Kontoghiorghes, E. J.  (2006).  Branch-and-bound
  algorithms for computing the best subset regression models.  Journal
  of Computational and Graphical Statistics, 15, 139-156.

}

\seealso{
  \code{\link{xsubset}}
  \code{\link{pollute}}
  \code{\link{house}}
}
